+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * 【extra  】     SenvenZExtractor改名为SevenZExtractor，增加getFirst、get方法
 * 【core   】     DateConverter修改返回java.util.Date而非DateTime（issue#I4BOAP@Gitee）
 * 【core   】     增加IterableIter、ComputeIter
+* 【core   】     CsvConfig增加disableComment方法（issue#1842@Github）
 
 ### 🐞Bug修复
 * 【http   】     修复HttpCookie设置cookies的方法，不符合RFC6265规范问题（pr#418@Gitee）